added departments page with table and all CRUD, Fixed in server bugs

+++ FAMA-RESTAPI/FAMA-RESTAPI/Models/departmentsBL.cs
@@ -8,10 +8,39 @@ namespace FAMA_RESTAPI.Models
     public class departmentsBL
     {
         famaDBEntities db = new famaDBEntities();
+        private static logsCheck log = new logsCheck();
 
         public IEnumerable<departments> getAllDepartments()
         {
             return db.departments;
+        }
+        public IEnumerable<departmentsExt> getAllDepartmentsAction(int userID)
+        {
+            famaDBEntities db = new famaDBEntities();
+            //add user action
+            log.addActionLog(userID);
+            //Check user actions
+            if (log.checkLogs(userID))
+            {
+                List<departmentsExt> depList = new List<departmentsExt>();
+                foreach (var dep in db.departments)
+                {
+                    departmentsExt department = new departmentsExt();
+                    department.ID = dep.ID;
+                    department.name = dep.name;
+                    department.managerID = dep.managerID;
+                    department.areAnyEmployees = false;
+                    foreach (var emp in db.employees)
+                    {
+                        if (emp.departmentID == dep.ID)
+                        {
+                            department.areAnyEmployees = true;
+                        }
+                    }
+                    depList.Add(department);
+                }
+                return depList;
+            } else { return null; }
         }
 
         public departments getDepartments(int id)
@@ -19,25 +48,40 @@ namespace FAMA_RESTAPI.Models
             return db.departments.Where(d => d.ID == id).First();
         }
 
-        public void postDepartments(departments dep)
+        public void postDepartments(int userID, departments dep)
         {
-            db.departments.Add(dep);
-            db.SaveChanges();
+            //add user action
+            log.addActionLog(userID);
+            //Check user actions
+            if (log.checkLogs(userID)) { 
+                db.departments.Add(dep);
+                db.SaveChanges();
+            }
         }
 
-        public void putDepartments(int id, departments dep)
+        public void putDepartments(int id, int userID, departments dep)
         {
-            var currentDepartment = db.departments.Where(d => d.ID == id).First();
-            currentDepartment.name = dep.name;
-            currentDepartment.managerID = dep.managerID;
-            db.SaveChanges();
+            //add user action
+            log.addActionLog(userID);
+            //Check user actions
+            if (log.checkLogs(userID)) { 
+                var currentDepartment = db.departments.Where(d => d.ID == id).First();
+                currentDepartment.name = dep.name;
+                currentDepartment.managerID = dep.managerID;
+                db.SaveChanges();
+            }
         }
 
-        public void deleteDepartments(int id)
+        public void deleteDepartments(int id, int userID)
         {
-            var currentDepartment = db.departments.Where(d => d.ID == id).First();
-            db.departments.Remove(currentDepartment);
-            db.SaveChanges();
+            //add user action
+            log.addActionLog(userID);
+            //Check user actions
+            if (log.checkLogs(userID)) { 
+                var currentDepartment = db.departments.Where(d => d.ID == id).First();
+                db.departments.Remove(currentDepartment);
+                db.SaveChanges();
+            }
         }
     }
 }
